changed voting type to interface; renamed folder to "interfaces";
