+++ src/app/interfaces/votings.ts
@@ -1,0 +1,9 @@
+export interface Voting {
+  id: number,
+  title: string,
+  voting_question: string,
+  description: string,
+  creator_id: string,
+  created_at: Date,
+  public: boolean
+}
