--- src/app/types/votings.ts
@@ -1,9 +1,0 @@
-export type Voting = {
-  "id": number,
-  "title": string,
-  "voting_question": string,
-  "description": string,
-  "creator_id": string,
-  "created_at": Date,
-  "public": boolean
-}
